refactor to use `Config`
